+++ ContactMechanics/Systems.py
@@ -197,14 +197,14 @@ class SystemBase(object, metaclass=abc.ABCMeta):
 
     def _reshape_bounds(self, lbounds=None, ubounds=None):
         if lbounds is not None and ubounds is not None:
-            ubounds = self.shape_minimisation_input(np.ma.filled(ubounds, np.inf))
-            lbounds = self.shape_minimisation_input(np.ma.filled(lbounds, -np.inf))
+            ubounds = (np.ma.filled(ubounds, np.inf)).ravel()
+            lbounds = (np.ma.filled(lbounds, -np.inf)).ravel()
             return optim.Bounds(lb=lbounds, ub=ubounds)
         elif lbounds is not None:
-            lbounds = self.shape_minimisation_input(np.ma.filled(lbounds, -np.inf))
+            lbounds = (np.ma.filled(lbounds, -np.inf)).ravel()
             return optim.Bounds(lb=lbounds)
         elif ubounds is not None:
-            ubounds = self.shape_minimisation_input(np.ma.filled(ubounds, np.inf))
+            ubounds = (np.ma.filled(ubounds, np.inf)).ravel()
             return optim.Bounds(ub=ubounds)
         else:
             return None
@@ -450,9 +450,11 @@ class NonSmoothContactSystem(SystemBase):
 
         # How to compute the contact area will actually depend on wether it is a primal or dual solver
         return (['energy',
-                 'substrate force', ],
+                 'force',
+                 ],
                 [self.energy,
-                 -self.reduction.sum(self.substrate.force), ])
+                 -self.reduction.sum(self.substrate.force),
+                 ])
 
     def evaluate(self, disp, offset, pot=True, forces=False, logger=None):
         """
@@ -754,7 +756,7 @@ class NonSmoothContactSystem(SystemBase):
 
         return result
 
-    def evaluate_dual(self, press, offset, forces=False):
+    def evaluate_dual(self, force, offset, logger=None):
         """
         Computes the energies and forces in the system for a given pressure field.
 
@@ -770,11 +772,11 @@ class NonSmoothContactSystem(SystemBase):
 
         Parameters
         ----------
-        press : array_like
-            The pressure field for which the displacement field is to be computed.
+        force : array_like
+            The force applied to each pixel of the substrate
         offset : float
             The offset value to be used in the computation.
-        forces : bool, optional
+        gradient : bool, optional
             If True, the forces in the system are also computed. Default is False.
 
         Returns
@@ -785,18 +787,49 @@ class NonSmoothContactSystem(SystemBase):
             Gradient, which is the difference between the displacement and the sum of the
             surface heights and the offset. If forces are not computed, the gradient is None.
         """
-        disp = self.substrate.evaluate_disp(-press)
-        if forces:
-            self.gradient = disp - self.surface.heights() - offset
-        else:
-            self.gradient = None
+        disp = self.substrate.evaluate_disp(-force)
 
-        self.energy = 1 / 2 * np.sum(press * disp) - np.sum(
-            press * (self.surface.heights() + offset))
+        self.gradient = disp - self.surface.heights() - offset
+
+        self.energy = 1 / 2 * np.sum(force * disp) - np.sum(
+            force * (self.surface.heights() + offset))
+
+        # Update state:
+        self.substrate.force = - force
+        self.substrate.disp = disp
+
+        if logger is not None:
+            tot_nb_grid_pts = np.prod(self.nb_grid_pts)
+            nb_rep = self.reduction.sum(
+                np.where(- self.substrate.force[
+                    self.substrate.local_topography_subdomain_slices]
+                         > 0, 1., 0.))
+
+            nb_att = self.reduction.sum(
+                np.where(- self.substrate.force[
+                    self.substrate.local_topography_subdomain_slices]
+                         < 0, 1., 0.))
+
+            proj_grad = self.gradient * (force > 0)
+            rms_grad = np.sqrt(self.reduction.sum(proj_grad ** 2) / tot_nb_grid_pts)
+            max_grad = self.reduction.max(np.abs(proj_grad))
+            logger.st(['energy',
+                       'force',
+                       'frac_rep_area',
+                       'frac_att_area',
+                       'max_residual',
+                       'rms_residual'],
+                      [self.energy,
+                       -self.reduction.sum(self.substrate.force),
+                       nb_rep / tot_nb_grid_pts,
+                       nb_att / tot_nb_grid_pts,
+                       max_grad,
+                       rms_grad,
+                       ])
 
         return self.energy, self.gradient
 
-    def dual_objective(self, offset, gradient=True):
+    def dual_objective(self, offset, gradient=True, logger=None):
         r"""
         Objective function to handle dual objective, i.e. the Legendre
         transformation from displacements as variable to pressures
@@ -814,7 +847,11 @@ class NonSmoothContactSystem(SystemBase):
         energy : float
             Value of total energy.
         gradient : array_like
-            Value of the gradient (array) or the value of gap (if gradient is True).
+            Value of the gradient (array), which corresponds to the value of gap..
+            The array is ravelled to a 1D arreay for compatibility with scipy solvers.
+            Note that for the nonperiodic substrate that uses padding, the shape of the array
+            corresponds to the unpadded domain, i.e. where the topography is defined.
+            That is, the shape of the array before ravelling is `substrate.topography_nb_subdomain_grid_pts`.`
 
         Notes
         -----
@@ -828,20 +865,19 @@ class NonSmoothContactSystem(SystemBase):
             \text{which is,} \\
             \text{gap} = \text{displacement} - \text{height} \\
         """
-        res = self.substrate.nb_domain_grid_pts
+        res = self.substrate.topography_nb_subdomain_grid_pts
         if gradient:
             def fun(pressure):
                 try:
                     self.evaluate_dual(
-                        pressure.reshape(res), offset, forces=True)
+                        pressure.reshape(res), offset, logger=logger)
                 except ValueError as err:
                     raise ValueError(
                         "{}: gap.shape: {}, res: {}".format(
                             err, pressure.shape, res))
                 return self.energy, self.gradient.ravel()
         else:
-            def fun(gap):
-                return self.evaluate(gap.reshape(res), forces=False)[0]
+            raise NotImplementedError
 
         return fun
 
@@ -863,11 +899,12 @@ class NonSmoothContactSystem(SystemBase):
             The hessian product, which is the result of applying the hessian matrix to the pressure vector.
         """
         inres = pressure.shape
-        res = self.substrate.nb_subdomain_grid_pts
+        res = self.substrate.topography_nb_subdomain_grid_pts
         hessp = self.substrate.evaluate_disp(-pressure.reshape(res))
         return hessp.reshape(inres)
 
-    def dual_minimize_proxy(self, offset, init_force=None, solver='ccg-without-restart', gtol=1e-8, maxiter=1000):
+    def dual_minimize_proxy(self, offset, init_force=None, solver='ccg-without-restart', gtol=1e-8, maxiter=1000,
+                            logger=None):
         """
         Convenience function for DUAL minimisation (pixel forces as variables).
         This function simplifies the process of solving the dual minimisation problem
@@ -879,6 +916,8 @@ class NonSmoothContactSystem(SystemBase):
             Determines the indentation depth.
         init_force : array_like, optional
             Initial guess for the force. If not provided, it defaults to None.
+            For Nonperiodic calculations, the array does not include the padding region,
+            i.e. it is of shape `substrate.topography_nb_subdomain_grid_pts`
         solver : str, optional
             The solver to be used for the minimisation. It can be one of
             'ccg-without-restart', 'ccg-with-restart', or 'l-bfgs-b'. If not provided,
@@ -906,28 +945,32 @@ class NonSmoothContactSystem(SystemBase):
         self.contact_zone = None
         self.init_force = init_force
 
+        if init_force is not None and init_force.shape != self.substrate.topography_nb_subdomain_grid_pts:
+            raise IncompatibleResolutionError
+
         lbounds = np.zeros(self.init_force.shape)
         bnds = self._reshape_bounds(lbounds, )
 
         if solver == 'ccg-without-restart':
             result = CCGWithoutRestart.constrained_conjugate_gradients(
-                self.dual_objective(offset, gradient=True),
+                self.dual_objective(offset, gradient=True, logger=logger),
                 self.dual_hessian_product, x0=init_force, gtol=gtol,
                 maxiter=maxiter)
         elif solver == 'ccg-with-restart':
             result = CCGWithRestart.constrained_conjugate_gradients(
-                self.dual_objective(offset, gradient=True),
+                self.dual_objective(offset, gradient=True, logger=logger),
                 self.dual_hessian_product, x0=init_force, gtol=gtol,
                 maxiter=maxiter)
         elif solver == 'l-bfgs-b':
             result = optim.minimize(
-                self.dual_objective(offset, gradient=True),
-                self.shape_minimisation_input(self.init_force),
+                self.dual_objective(offset, gradient=True, logger=logger),
+                self.init_force.ravel(),
                 method='L-BFGS-B', jac=True,
                 bounds=bnds,
                 options=dict(gtol=gtol, ftol=1e-20))
 
         if result.success:
+            # TODO: I think I need to call substrate.compute , so that substrate.energy is computed
             self.offset = offset
             self.gap = result.jac
             self.force = self.substrate.force = result.x