+++ test/test_primal_dual_objective.py
@@ -5,6 +5,7 @@ import scipy.optimize as optim
 
 import numpy as np
 import pytest
+from ContactMechanics.Tools.Logger import screen, Logger
 
 
 # import matplotlib.pyplot as plt
@@ -82,7 +83,7 @@ def test_dual_obj(s):
     init_force = substrate.evaluate_force(disp)
 
     res = system.dual_minimize_proxy(offset, init_force=init_force,
-                                     solver='ccg-without-restart', )
+                                     solver='ccg-without-restart', logger=Logger('test_log.log'))
 
     assert res.success
     CA_bug = res.x.reshape((nx, ny)) > 0  # Contact area
@@ -92,7 +93,7 @@ def test_dual_obj(s):
     gap_bug = gap_bug.reshape((nx, ny))
 
     res = system.dual_minimize_proxy(offset, init_force=init_force,
-                                     solver='ccg-with-restart', )
+                                     solver='ccg-with-restart', logger=screen)
 
     assert res.success, res.message
     CA_pk = res.x.reshape((nx, ny)) > 0  # Contact are
@@ -102,7 +103,7 @@ def test_dual_obj(s):
     gap_pk = gap_pk.reshape((nx, ny))
 
     res = system.dual_minimize_proxy(offset, init_force=init_force,
-                                     solver='l-bfgs-b', )
+                                     solver='l-bfgs-b', logger=screen)
 
     assert res.success, res.message
     CA_lbfgsb = res.x.reshape((nx, ny)) > 0  # Contact area
@@ -117,6 +118,32 @@ def test_dual_obj(s):
     np.testing.assert_allclose(CA_pk, CA_lbfgsb, atol=1e-5)
     np.testing.assert_allclose(gap_pk, gap_lbfgsb, atol=1e-5)
     np.testing.assert_allclose(force_pk, force_lbfgsb, atol=1e-5)
+
+
+@pytest.mark.parametrize("solver", {'ccg-without-restart', 'ccg-with-restart', 'l-bfgs-b'})
+def test_dual_minimize_proxy_nonperiodic(solver):
+    nx, ny = 128, 128
+    sx = sy = 1.
+    R = 10.
+    Es = 50.
+
+    surface = make_sphere(R, (nx, ny), (sx, sy), kind="paraboloid")
+    substrate = Solid.FreeFFTElasticHalfSpace((nx, ny), young=Es,
+                                              physical_sizes=(sx, sy))
+
+    offset = 0.005
+
+    init_gap = np.zeros((nx, ny))
+    disp = np.zeros((2 * nx, 2 * ny))
+    disp[substrate.topography_subdomain_slices] = init_gap + surface.heights() + offset
+    disp[disp < 0] = 0
+    init_force = substrate.evaluate_force(disp)[substrate.topography_subdomain_slices]
+
+    system = Solid.Systems.NonSmoothContactSystem(substrate, surface)
+
+    res = system.dual_minimize_proxy(offset, init_force=init_force,
+                                     solver=solver, logger=Logger('test_log.log'))
+    assert res.success
 
 
 @pytest.mark.parametrize("s", (1., 2.))